requirements/test.txt
-rbase.txt
pytest==7.2.0
pytest-mock==3.10.0
pytest-cov==4.0.0
pytest-sqlalchemy==0.2.1
